@@ -2,9 +2,9 @@ import React, { useCallback } from 'react';
 import './App.css';
 import {InArrayProps, Todolist} from "./Todolist";
 import {AddItemForm} from "./Components/AddItemForm";
-import {AppBar, Container, Grid, IconButton, Paper, Toolbar, Typography} from '@material-ui/core';
-import {Menu} from '@material-ui/icons';
-import {Button} from "@material-ui/core";
+
+
+
 import {
     AddTodolistAC,
     ChangeTodolistAC,
@@ -21,6 +21,8 @@ import {
 } from "./state/tasks-reducer";
 import {useDispatch, useSelector} from "react-redux";
 import {AppRootState} from "./state/store";
+import {AppBar, Button, Container, Grid, IconButton, Paper, Toolbar, Typography} from '@mui/material';
+import {Menu} from "@mui/icons-material";
 //
 export type FilterValuesType = 'all' | 'active' | 'completed'
 export type TodolistType = {
